agent/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
openai>=1.30.0
websockets>=13.0
httpx>=0.27.0
pydantic>=2.0
python-dotenv>=1.0.0
duckduckgo-search>=6.0.0
playwright>=1.40.0
pyyaml>=6.0
